Clickstream/requirements.txt
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
findspark==2.0.1
fonttools==4.38.0
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pure-sasl==0.6.2
py4j==0.10.9.5
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.42.1
PyHive==0.6.5
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.2
PyScreeze==0.1.28
pyspark==3.3.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.8
requests==2.25.1
sasl==0.3.1
schedule==1.1.0
seaborn==0.12.1
SecretStorage==3.3.1
Send2Trash==1.8.0
simple-colors==0.1.5
six==1.16.0
snakebite-py3==3.0.5
sniffio==1.3.0
soupsieve==2.3.2.post1
ssh-import-id==5.11
stack-data==0.6.2
systemd-python==234
terminado==0.17.0
terminator==2.1.1
thrift==0.16.0
thrift-sasl==0.4.3
tinycss2==1.2.1
tornado==6.2
traitlets==5.6.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
xdg==5
xkit==0.0.0
zipp==1.0.0